Accurately model Asset.media_url
This was a URLField but because Django's model validation had never been executed nobody had noticed
that the values were never valid URLs, only path prefixes.

There's a separate question about whether we should even have this field at all since we're using a rigid sequence numbering system, and it should be refactored to be project-relative so we don't need to update the data if the project or campaign portions of the URL change.