Enable django-storages
This avoids having so many branchy code paths which
require care to test (a separate bug caused it to often use S3 even when the settings file appeared to
be using local storage) and instead uses the standard
Django storage system, allowing the switch between local and network storage to be a settings change.

This commit also introduces a named ASSET_STORAGE backend definition in concordia.storage to make it easier to use separate storage pools in the future.