WIP: asset view cleanup
This is starting the process of removing the internal 
self-calls which break production and cause significant
performance problems everywhere else.

The new model will be that the asset detail view is a normal
detail view and the user inputs will be AJAX POSTed to the
API directly to avoid the need to have a ton of code which
copies cookies & tokens around and to provide the richer UI
which we want for allowing users to select the next action
after a successful submission.