Transcription view state changes
Now the status is updated as users work. When a transcription
is saved it will be recorded as superseding the previous
transcription, if any, and the Submit for Review button will
only be enabled if the transcription has not been modified
since the last time it was saved.

Successfully submitting a transcription for review will mark
it as “Submitted for Review” in the page and disable the
editing controls.