Basic implementation of the next-transcribable button
This adds a basic redirect to find the next asset which has
not yet been submitted and for which the user can get a
reservation.

It handles failure to get a reservation by redirecting to
the project page.

It does not yet attempt to favor assets which are in the
same item as the last asset which the user was working on.

See #286
See #372