Rebuild CAPTCHA implementation
This now uses a fully AJAX approach which will only prompt
the user when the underlying view being called returns an
HTTP 401 error.

This adds a @validate_anonymous_captcha view decorator which
either processes the view as expected or returns an HTTP 401
with the CAPTCHA key & challenge image URL. 

The AJAX handler for form submissions automatically displays
the CAPTCHA modal and the error handler will update the 
image as many times as needed before successful completion.

It does not currently repeat the operation which the user 
was attempting since that would involve preserving the state
for whether the operation being attempted was a save or a
submission. We can add that after the UX is accepted.