Update contributor report user breakdown
We consider transcription the core feature for this report so we'll
report that number instead of the number of users who've helped tag
items.

Transcription and reviewing are separate activities and they're now
listed independently to avoid a fairly big subquery needed to calculate
the unique values across both fields because Django's subquery support
doesn't support .count().