Replace code copied from django-passwords
The complexity validator was copied from @dstufft’s 
django-passwords project in violation of the license:

https://github.com/dstufft/django-passwords/blob/master/LICENSE

Based on 56f08d282e7a027da0739e3279209609e0dbee48
this might have been an attempt to make it 
compatible with the official Django password 
validation API added in Django 1.9. This commit
changes that to use django-passwords directly and
just wraps the interface slightly.