indexes: more efficient database queries
Using prefetch_related to avoid on-demand queries
for the nested object hierarchy, which reduces the
total runtime by roughly a factor of 4, which will
hopefully be enough until @mjl’s upstream 
pull-request is merged:

https://github.com/sabricot/django-elasticsearch-dsl/pull/154

queryset_pagination seemed like a potential area
for improvement but in practice was not faster.

Using selected_related was faster than nothing but
about half as fast as prefetch_related since it
retrieves so much duplicate data.