Refactor AssetViewer update signature
Now the AssetViewer update() method expects to
receive an object with the same values as before
and a new editable: object containing the two
canEdit: and reason: properties used elsewhere.

This also changes the action handler to check the
editable status after each update so the views
simply receive those values and do not need logic
to calculate them.