Expose reservation token to action app
We need to coordinate having both the list views
and the WebSocket updates do per-user can-edit 
flags before we can get rid of this because we 
need to filter the asset list for assets which can
possibly be edited (i.e. asset.reservation token
is null) and those which are actively in use by
someone else.