canEditAsset now tests asset editability
This will flag an asset as unable to be edited if
there is a known reservation token which does not
match our reservation token. This does not 
guarantee that it can be edited so we still need
to get the asset reservation when selected but it
does mean that the asset list can reliably badge
things which are known not to be possible to work
on.