tests: prevent spurious rate-limiting failures
The tests which depend on query codes may fail due
to the extra checks which the rate-limiting system
uses.

This is set in settings_test but now is forced off
for all of the view tests to avoid needing to have
developers remember to disable it before running
the test suite.