CloudFormation: use ARN for ExecutionRoleArn
The CFN validation is lax and only checks for strings but the actual ECS service requires these to be full ARNs. As of v0.20.0 cfn-lint reports this, as it did after I upgraded it in #943.

As a followup to this commit I’m going to have CFN create that role so we won’t have the external dependency.