Transcription View: avoid updates for unmodified sources
This avoids the problem in #949 where the update
triggered by the reservation system would cause
the transcription view to replace any unsaved text
with the original starting text. Now the update
method will not replace the text when the upstream
value has not changed.

This still requires testing around the case where
someone may not have had a lock when they first
clicked on the item to confirm that we correctly
will reset it when the lock is obtained.