Refactor setup to use switchMode
This ensures that we have the same path for the initial setup and subsequent mode transitions, which avoids hitting a CSS problem when the initial app element did not have the `data-mode` attribute set, which is what the CSS uses to set `display: flex` (vs. `none`) on the transcription & review columns.  

This became a problem when the introduction of the persistent URL states meant it was possible to start in transcription mode without having passed through review mode first.